Add an action to set UDFs as not required in the data model for a project. The actions.csv file should be read to get the field names to set. This should ONLY work on UDFs. If TDFs or system fields are specified they should be ignored.
Example syntax:
ALLSTD,SetFieldToNotRequired,<DBTableName>,<fieldname>,,
Example:
ALLSTD,SetFieldToNotRequired,Bug,Team,,
This would set the Team field to Not Required, but only if it is a UDF (e.g, the DB column name starts with "BG_USER_").

In addition to reading specific field names from actions.csv, provide the ability to use an asterick ("*") wildcard. This would direct the tool to set all fields in the specified project table to Not Required.
Example:
ALLSTD,SetFieldToNotRequired,Bug,*,,
In this example all UDFs in the Bug table would be set to Not Required.